NUMBERS-159: move classes from commons-numbers-arrays to more appropriate modules
