+++ commons-numbers-angle/src/test/java/org/apache/commons/numbers/angle/CosAngleTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.commons.numbers.arrays;
+package org.apache.commons.numbers.angle;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
