+++ commons-numbers-core/src/main/java/org/apache/commons/numbers/core/Norms.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.commons.numbers.arrays;
+package org.apache.commons.numbers.core;
 
 /** Class providing methods to compute various norm values.
  *
