+++ commons-numbers-core/src/test/java/org/apache/commons/numbers/core/DoubleTestUtils.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.commons.numbers.arrays;
+package org.apache.commons.numbers.core;
 
 import org.apache.commons.rng.UniformRandomProvider;
