+++ commons-numbers-examples/examples-jmh/src/test/java/org/apache/commons/numbers/examples/jmh/core/EuclideanNormAccuracyTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.commons.numbers.examples.jmh.arrays;
+package org.apache.commons.numbers.examples.jmh.core;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
